Move currency list to new currency module
