+++ apps/api/src/app/app.controller.ts
@@ -1,3 +1,4 @@
+import { CurrencyService } from '@ghostfolio/api/services/currency/currency.service';
 import { ExchangeRateDataService } from '@ghostfolio/api/services/exchange-rate-data/exchange-rate-data.service';
 
 import { Controller } from '@nestjs/common';
@@ -5,6 +6,7 @@ import { Controller } from '@nestjs/common';
 @Controller()
 export class AppController {
   public constructor(
+    private readonly currencyService: CurrencyService,
     private readonly exchangeRateDataService: ExchangeRateDataService
   ) {
     this.initialize();
@@ -12,6 +14,7 @@ export class AppController {
 
   private async initialize() {
     try {
+      await this.currencyService.initialize();
       await this.exchangeRateDataService.initialize();
     } catch {}
   }
