+++ apps/api/src/app/portfolio/calculator/twr/portfolio-calculator-fee.spec.ts
@@ -63,12 +63,7 @@ describe('PortfolioCalculator', () => {
 
     currentRateService = new CurrentRateService(null, null, null, null);
 
-    exchangeRateDataService = new ExchangeRateDataService(
-      null,
-      null,
-      null,
-      null
-    );
+    exchangeRateDataService = new ExchangeRateDataService(null, null, null);
 
     portfolioSnapshotService = new PortfolioSnapshotService(null);
