+++ apps/api/src/app/portfolio/calculator/twr/portfolio-calculator-googl-buy.spec.ts
@@ -76,12 +76,7 @@ describe('PortfolioCalculator', () => {
 
     currentRateService = new CurrentRateService(null, null, null, null);
 
-    exchangeRateDataService = new ExchangeRateDataService(
-      null,
-      null,
-      null,
-      null
-    );
+    exchangeRateDataService = new ExchangeRateDataService(null, null, null);
 
     portfolioSnapshotService = new PortfolioSnapshotService(null);
