+++ apps/api/src/app/portfolio/calculator/twr/portfolio-calculator-novn-buy-and-sell.spec.ts
@@ -77,12 +77,7 @@ describe('PortfolioCalculator', () => {
 
     currentRateService = new CurrentRateService(null, null, null, null);
 
-    exchangeRateDataService = new ExchangeRateDataService(
-      null,
-      null,
-      null,
-      null
-    );
+    exchangeRateDataService = new ExchangeRateDataService(null, null, null);
 
     portfolioSnapshotService = new PortfolioSnapshotService(null);
